functions/requirements.txt
aiofiles==25.1.0
azure-functions==1.24.0
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-storage==24.0.0
azure-storage-blob[aio]==12.27.1
python-dotenv==1.2.1
